Tests work again.
